Actually, let's move engine_loop to another file.
