+++ meson.build
@@ -182,6 +182,7 @@ common_files += [
 
 files += [
   'src/engine.cc',
+  'src/engine_loop.cc',
   'src/mcts/params.cc',
   'src/mcts/search.cc',
   'src/mcts/stoppers/alphazero.cc',
